+++ CHANGELOG.md
@@ -74,6 +74,7 @@
 - Celsius parser: added new export format.
 - Conversion tool: added parser for CEX.IO exchange.
 - Setup.py fixed for Python 3.11.
+- Binance parser: added re-named operations for Statements. ([#292](https://github.com/BittyTax/BittyTax/issues/292))
 ### Changed
 - Binance parser: performance improvements for large data sets.
 - Coinbase Pro parser: performance improvements for large data sets.